requirements.txt
boto3==1.17.45
botocore==1.20.112
jmespath==0.10.0
python-dateutil==2.8.2
s3transfer==0.3.7
six==1.16.0
slack_sdk==3.11
urllib3==1.26.6